Change TTL from uint16 to int for tcp routes
- When a route with negative TTL value is created, the error
  returned is about not being able to render JSON. Instead a meaningful
  error should be returned

[#119118969]

+++ cmd/tcp-emitter/main.go
@@ -188,7 +188,7 @@ func main() {
 		os.Exit(1)
 	}
 
-	emitter := emitter.NewEmitter(logger, routingAPIClient, uaaClient, uint16(routeTTL.Seconds()))
+	emitter := emitter.NewEmitter(logger, routingAPIClient, uaaClient, int(routeTTL.Seconds()))
 	routingTable := schema.NewTable(logger, nil)
 	routingTableHandler := routing_table.NewRoutingTableHandler(logger, routingTable, emitter, bbsClient)
 	syncChannel := make(chan struct{})
